changed LargeDataStores to store size as a long.  Fixed broken unit test and removed prototype code that was accidentally committed.
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@1090 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/common/core/assembly/ace/LargeAceFileDataStore.java
@@ -35,7 +35,7 @@ import org.jcvi.common.core.util.iter.CloseableIterator;
 public final class LargeAceFileDataStore extends AbstractDataStore<AceContig> implements AceContigDataStore{
 
 	private final File aceFile;
-	private Integer numberOfContigs = null;
+	private Long numberOfContigs = null;
 	private final DataStoreFilter contigIdFilter;
 	
 	/**
@@ -165,10 +165,10 @@ public final class LargeAceFileDataStore extends AbstractDataStore<AceContig> im
 
 	private class SizeVisitor implements AceFileVisitor{
 
-		private int size=0;
+		private long size=0;
 		
 		
-		public int getSize() {
+		public long getSize() {
 			return size;
 		}
